requirements.txt
aifunctions @ file:///C:/Users/kryardley/OneDrive%20-%20Church%20of%20Jesus%20Christ/Documents/pdf_extract/SimpleAzureSearch-template/aifunctions
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
audioread==3.0.1
azure-cognitiveservices-speech==1.40.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-search==9.1.0
azure-search-documents==11.5.1
blinker==1.8.2
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.5
decorator==4.4.2
distro==1.9.0
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.3
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
JPype1==1.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
nest-asyncio==1.6.0
noisereduce==3.0.2
numba==0.60.0
numpy==2.0.2
openai==1.45.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pdfquery==0.4.3
pillow==10.4.0
platformdirs==4.3.2
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==2.10.5
pypdfium2==4.30.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.2.0
requests==2.32.3
roman==4.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
stack-data==0.6.3
tabula-py==2.9.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.4
